Simplified Taskbar COM

+++ src/BenchmarkDotNet/Running/BenchmarkRunnerClean.cs
@@ -658,7 +658,7 @@ namespace BenchmarkDotNet.Running
             {
                 Console.Title = $"{benchmarksToRunCount}/{totalBenchmarkCount} Remaining - {(int)fromNow.TotalHours}h {fromNow.Minutes}m to finish";
             }
-            taskbarProgress.SetProgress(executedBenchmarkCount, totalBenchmarkCount);
+            taskbarProgress.SetProgress((ulong) executedBenchmarkCount, (ulong) totalBenchmarkCount);
         }
 
         private static TimeSpan GetEstimatedFinishTime(in StartedClock runsChronometer, int benchmarksToRunCount, int executedBenchmarkCount)
